pythonlogin/requirements.txt
Flask
flask-session
timedelta
psycopg2-binary
gunicorn
